backend/requirements.txt

flask==2.0.2
flask-cors==3.0.10
pymongo==4.1.1
python-dotenv==0.19.2
